The ID of particles is now a 'long' and not an 'int' anymore for compatibility with GADGET and because we will use more than 10^9 particles anyway.
Former-commit-id: 4e086d540cae2f68080a77e055a146d5b35f130a

+++ src/part.h
@@ -37,7 +37,7 @@ struct part {
     float mass;
     
     /* Particle ID. */
-    int id;
+    long id;
     
     /* Particle position. */
     double x[3];
